Update folder structure (#84)
This PR updates the folder structure to allow a cleaner separation of
functionality:
1. We introduce a `torch` folder, where RL implementations like the
different SAC trainers are stored.
2. We provide an `ocp` folder, where interfaces to OCP solvers like
Acados are stored.
3. The `autograd` folder contains all the code necessary for wrapping
code such that it can be used for `backward` and `forward` passes in Jax
or PyTorch.

We update the test folder structure accordingly.
