+++ leap_c/examples/pointmass/mpc.py
@@ -9,7 +9,7 @@ from leap_c.examples.util import (
     find_param_in_p_or_p_global,
     translate_learnable_param_to_p_global,
 )
-from leap_c.acados.mpc import Mpc
+from leap_c.ocp.acados.mpc import Mpc
 
 
 class PointMassMPC(Mpc):
