+++ leap_c/examples/quadrotor/task.py
@@ -7,11 +7,11 @@ from gymnasium import spaces
 
 from leap_c.examples.quadrotor.env import QuadrotorStop
 from leap_c.examples.quadrotor.mpc import QuadrotorMpc
-from leap_c.acados.layer import MpcSolutionModule
+from leap_c.ocp.acados.layer import MpcSolutionModule
 from leap_c.registry import register_task
 from leap_c.task import Task
 
-from leap_c.acados.mpc import MpcInput, MpcParameter
+from leap_c.ocp.acados.mpc import MpcInput, MpcParameter
 
 
 @register_task("quadrotor_terminal")
