--- leap_c/rl/__init__.py
@@ -1,3 +1,0 @@
-import leap_c.rl.sac  # noqa: F401, F403
-import leap_c.rl.sac_fop  # noqa: F401, F403
-import leap_c.rl.sac_zop  # noqa: F401, F403
