+++ leap_c/run.py
@@ -5,7 +5,6 @@ from pathlib import Path
 
 import leap_c.examples  # noqa: F401
 from leap_c.registry import create_default_cfg, create_task, create_trainer
-import leap_c.rl
 from leap_c.trainer import BaseConfig
 from leap_c.utils.cfg import cfg_as_python
 from leap_c.utils.git import log_git_hash_and_diff
