+++ leap_c/torch/rl/sac_fop.py
@@ -10,14 +10,14 @@ import numpy as np
 import torch
 import torch.nn as nn
 
-from leap_c.acados.mpc import MpcBatchedState
-from leap_c.nn.gaussian import SquashedGaussian, BoundedTransform
-from leap_c.nn.mlp import MLP, MlpConfig
-from leap_c.acados.layer import MpcSolutionModule
+from leap_c.ocp.acados.mpc import MpcBatchedState
+from leap_c.torch.nn.gaussian import SquashedGaussian, BoundedTransform
+from leap_c.torch.nn.mlp import MLP, MlpConfig
+from leap_c.ocp.acados.layer import MpcSolutionModule
 from leap_c.registry import register_trainer
-from leap_c.rl.buffer import ReplayBuffer
-from leap_c.rl.sac import SacBaseConfig, SacCritic
-from leap_c.rl.utils import soft_target_update
+from leap_c.torch.rl.buffer import ReplayBuffer
+from leap_c.torch.rl.sac import SacBaseConfig, SacCritic
+from leap_c.torch.rl.utils import soft_target_update
 from leap_c.task import Task
 from leap_c.trainer import Trainer
