+++ leap_c/torch/rl/sac_zop.py
@@ -1,82 +1,28 @@
-from dataclasses import dataclass, field
+"""Provides a trainer for a Soft Actor-Critic algorithm that uses a differentiable MPC
+layer for the policy network."""
+
 from pathlib import Path
-from typing import Iterator
+from typing import Any, Iterator, NamedTuple
 
 import gymnasium as gym
 import numpy as np
 import torch
 import torch.nn as nn
 
-from leap_c.utils.logger import LoggerConfig
-from leap_c.nn.gaussian import SquashedGaussian
-from leap_c.nn.mlp import MLP, MlpConfig
-from leap_c.nn.utils import min_max_scaling
+from leap_c.ocp.acados.mpc import MpcBatchedState
+from leap_c.torch.nn.gaussian import SquashedGaussian
+from leap_c.torch.nn.mlp import MLP, MlpConfig
+from leap_c.ocp.acados.layer import MpcSolutionModule
+from leap_c.torch.nn.scale import min_max_scaling
 from leap_c.registry import register_trainer
-from leap_c.rl.buffer import ReplayBuffer
-from leap_c.rl.utils import soft_target_update
+from leap_c.torch.rl.buffer import ReplayBuffer
+from leap_c.torch.rl.utils import soft_target_update
+from leap_c.torch.rl.sac import SacBaseConfig
 from leap_c.task import Task
-from leap_c.trainer import BaseConfig, TrainConfig, Trainer, ValConfig
+from leap_c.trainer import Trainer
 
 
-@dataclass(kw_only=True)
-class SacAlgorithmConfig:
-    """Contains the necessary information for a SacTrainer.
-
-    Attributes:
-        critic_mlp: The configuration for the critic networks.
-        actor_mlp: The configuration for the actor network.
-        batch_size: The batch size for training.
-        buffer_size: The size of the replay buffer.
-        gamma: The discount factor.
-        tau: The soft update factor.
-        soft_update_freq: The frequency of soft updates.
-        lr_q: The learning rate for the Q networks.
-        lr_pi: The learning rate for the policy network.
-        lr_alpha: The learning rate for the temperature parameter.
-        init_alpha: The initial temperature parameter.
-        target_entropy: The minimum target entropy for the policy. If None, it
-            is set automatically depending on dimensions of the action space.
-        entropy_reward_bonus: Whether to add an entropy bonus to the reward.
-        num_critics: The number of critic networks.
-        report_loss_freq: The frequency of reporting the loss.
-        update_freq: The frequency of updating the networks.
-    """
-
-    critic_mlp: MlpConfig = field(default_factory=MlpConfig)
-    actor_mlp: MlpConfig = field(default_factory=MlpConfig)
-    batch_size: int = 64
-    buffer_size: int = 1000000
-    gamma: float = 0.99
-    tau: float = 0.005
-    soft_update_freq: int = 1
-    lr_q: float = 1e-4
-    lr_pi: float = 1e-4
-    lr_alpha: float | None = 1e-3
-    init_alpha: float = 0.01
-    target_entropy: float | None = None
-    entropy_reward_bonus: bool = True
-    num_critics: int = 2
-    report_loss_freq: int = 100
-    update_freq: int = 4
-
-
-@dataclass(kw_only=True)
-class SacBaseConfig(BaseConfig):
-    """Contains the necessary information for a Trainer.
-
-    Attributes:
-        sac: The Sac algorithm configuration.
-        train: The training configuration.
-        val: The validation configuration.
-        log: The logging configuration.
-        seed: The seed for the trainer.
-    """
-
-    sac: SacAlgorithmConfig = field(default_factory=SacAlgorithmConfig)
-    train: TrainConfig = field(default_factory=TrainConfig)
-    val: ValConfig = field(default_factory=ValConfig)
-    log: LoggerConfig = field(default_factory=LoggerConfig)
-    seed: int = 0
+NUM_THREADS_ACADOS_BATCH = 4
 
 
 class SacCritic(nn.Module):
@@ -89,7 +35,7 @@ class SacCritic(nn.Module):
     ):
         super().__init__()
 
-        action_dim = env.action_space.shape[0]  # type: ignore
+        action_dim = task.param_space.shape[0]  # type: ignore
 
         self.extractor = nn.ModuleList(
             [task.create_extractor(env) for _ in range(num_critics)]
@@ -104,42 +50,81 @@ class SacCritic(nn.Module):
                 for qe in self.extractor
             ]
         )
-        self.action_space = env.action_space
+        self.param_space = task.param_space
 
-    def forward(self, x: torch.Tensor, a: torch.Tensor):
-        a_norm = min_max_scaling(a, self.action_space)  # type: ignore
-        return [mlp(qe(x), a_norm) for qe, mlp in zip(self.extractor, self.mlp)]
+    def forward(self, x: torch.Tensor, p: torch.Tensor):
+        p_norm = min_max_scaling(p, self.param_space)  # type: ignore
+        return [mlp(qe(x), p_norm) for qe, mlp in zip(self.extractor, self.mlp)]
 
 
-class SacActor(nn.Module):
-    scale: torch.Tensor
-    loc: torch.Tensor
+class SacZopActorOutput(NamedTuple):
+    param: torch.Tensor
+    log_prob: torch.Tensor
+    stats: dict[str, float]
+    action: torch.Tensor | None = None
+    status: torch.Tensor | None = None
+    state_solution: MpcBatchedState | None = None
 
-    def __init__(self, task, env, mlp_cfg: MlpConfig):
+
+class MpcSacActor(nn.Module):
+    def __init__(
+        self,
+        task: Task,
+        env: gym.Env,
+        mlp_cfg: MlpConfig,
+    ):
         super().__init__()
 
-        self.extractor = task.create_extractor(env)
-        action_dim = env.action_space.shape[0]  # type: ignore
+        param_space = task.param_space
 
+        self.extractor = task.create_extractor(env)
         self.mlp = MLP(
             input_sizes=self.extractor.output_size,
-            output_sizes=(action_dim, action_dim),  # type: ignore
+            output_sizes=(param_space.shape[0], param_space.shape[0]),  # type:ignore
             mlp_cfg=mlp_cfg,
         )
 
-        self.squashed_gaussian = SquashedGaussian(env.action_space)
+        self.mpc: MpcSolutionModule = task.mpc  # type:ignore
+        self.prepare_mpc_input = task.prepare_mpc_input
 
-    def forward(self, x: torch.Tensor, deterministic=False):
-        e = self.extractor(x)
+        self.squashed_gaussian = SquashedGaussian(param_space)  # type:ignore
+
+    def forward(
+        self,
+        obs,
+        mpc_state: None | MpcBatchedState,
+        deterministic: bool = False,
+        only_param: bool = False,
+    ) -> SacZopActorOutput:
+        e = self.extractor(obs)
         mean, log_std = self.mlp(e)
 
-        action, log_prob, stats = self.squashed_gaussian(mean, log_std, deterministic)
+        param, log_prob, gauss_stats = self.squashed_gaussian(
+            mean, log_std, deterministic
+        )
 
-        return action, log_prob, stats
+        if only_param:
+            return SacZopActorOutput(param, log_prob, gauss_stats)
+
+        mpc_input = self.prepare_mpc_input(obs, param)
+
+        # TODO: We have to catch and probably replace the state_solution somewhere,
+        #       if its not a converged solution
+        with torch.no_grad():
+            mpc_output, state_solution, mpc_stats = self.mpc(mpc_input, mpc_state)
+
+        return SacZopActorOutput(
+            param,
+            log_prob,
+            {**gauss_stats, **mpc_stats},
+            mpc_output.u0,
+            mpc_output.status,
+            state_solution,  # type: ignore
+        )
 
 
-@register_trainer("sac", SacBaseConfig())
-class SacTrainer(Trainer):
+@register_trainer("sac_zop", SacBaseConfig())
+class SacZopTrainer(Trainer):
     cfg: SacBaseConfig
 
     def __init__(
@@ -164,14 +149,17 @@ class SacTrainer(Trainer):
         self.q_target.load_state_dict(self.q.state_dict())
         self.q_optim = torch.optim.Adam(self.q.parameters(), lr=cfg.sac.lr_q)
 
-        self.pi = SacActor(task, self.train_env, cfg.sac.actor_mlp)  # type: ignore
+        self.pi = MpcSacActor(task, self.train_env, cfg.sac.actor_mlp)
+        self.pi.mpc.mpc.num_threads_batch_methods = NUM_THREADS_ACADOS_BATCH
         self.pi_optim = torch.optim.Adam(self.pi.parameters(), lr=cfg.sac.lr_pi)
 
-        self.log_alpha = nn.Parameter(torch.tensor(self.cfg.sac.init_alpha).log())  # type: ignore
+        self.log_alpha = nn.Parameter(torch.tensor(cfg.sac.init_alpha).log())  # type: ignore
 
-        if self.cfg.sac.lr_alpha is not None:
+        action_dim = np.prod(self.train_env.action_space.shape)  # type: ignore
+        param_dim = np.prod(task.param_space.shape)  # type: ignore
+        self.entropy_norm = param_dim / action_dim
+        if cfg.sac.lr_alpha is not None:
             self.alpha_optim = torch.optim.Adam([self.log_alpha], lr=cfg.sac.lr_alpha)  # type: ignore
-            action_dim = np.prod(self.train_env.action_space.shape)  # type: ignore
             self.target_entropy = (
                 -action_dim
                 if cfg.sac.target_entropy is None
@@ -185,15 +173,24 @@ class SacTrainer(Trainer):
 
     def train_loop(self) -> Iterator[int]:
         is_terminated = is_truncated = True
+        policy_state = None
+        obs = None
 
         while True:
             if is_terminated or is_truncated:
                 obs, _ = self.train_env.reset()
+                policy_state = None
                 is_terminated = is_truncated = False
 
-            action, _, stats = self.act(obs)  # type: ignore
-            self.report_stats("train_trajectory", {"action": action}, verbose=True)
-            self.report_stats("train_policy_rollout", stats, verbose=True)  # type: ignore
+            obs_batched = self.task.collate([obs], device=self.device)
+
+            with torch.no_grad():
+                pi_output = self.pi(obs_batched, policy_state, deterministic=False)
+                action = pi_output.action.cpu().numpy()[0]
+                param = pi_output.param.cpu().numpy()[0]
+
+            self.report_stats("train_trajectory", {"action": action, "param": param}, verbose=True)
+            self.report_stats("train_policy_rollout", pi_output.stats, verbose=True)
 
             obs_prime, reward, is_terminated, is_truncated, info = self.train_env.step(
                 action
@@ -205,10 +202,18 @@ class SacTrainer(Trainer):
                     stats.update(info["task"])
                 self.report_stats("train", info["episode"])
 
-            # TODO (Jasper): Add is_truncated to buffer.
-            self.buffer.put((obs, action, reward, obs_prime, is_terminated))  # type: ignore
+            self.buffer.put(
+                (
+                    obs,
+                    param,
+                    reward,
+                    obs_prime,
+                    is_terminated,
+                )
+            )  # type: ignore
 
             obs = obs_prime
+            policy_state = pi_output.state_solution
 
             if (
                 self.state.step >= self.cfg.train.start
@@ -219,7 +224,9 @@ class SacTrainer(Trainer):
                 o, a, r, o_prime, te = self.buffer.sample(self.cfg.sac.batch_size)
 
                 # sample action
-                a_pi, log_p, _ = self.pi(o)
+                pi_o = self.pi(o, None, only_param=True)
+                a_pi = pi_o.param
+                log_p = pi_o.log_prob / self.entropy_norm
 
                 # update temperature
                 if self.alpha_optim is not None:
@@ -233,15 +240,15 @@ class SacTrainer(Trainer):
                 # update critic
                 alpha = self.log_alpha.exp().item()
                 with torch.no_grad():
-                    a_pi_prime, log_p_prime, _ = self.pi(o_prime)
-                    q_target = torch.cat(self.q_target(o_prime, a_pi_prime), dim=1)
+                    pi_o_prime = self.pi(o_prime, None, only_param=True)
+                    q_target = torch.cat(
+                        self.q_target(o_prime, pi_o_prime.param), dim=1
+                    )
                     q_target = torch.min(q_target, dim=1, keepdim=True).values
 
                     # add entropy
-                    q_target = (
-                        q_target
-                        - alpha * log_p_prime * self.cfg.sac.entropy_reward_bonus
-                    )
+                    factor = self.cfg.sac.entropy_reward_bonus / self.entropy_norm
+                    q_target = q_target - alpha * pi_o_prime.log_prob * factor
 
                     target = (
                         r[:, None] + self.cfg.sac.gamma * (1 - te[:, None]) * q_target
@@ -275,17 +282,21 @@ class SacTrainer(Trainer):
                     "q_target": target.mean().item(),
                     "entropy": -log_p.mean().item(),
                 }
-                self.report_stats("loss", loss_stats)
+                self.report_stats("loss", loss_stats, verbose=True)
 
             yield 1
 
     def act(
         self, obs, deterministic: bool = False, state=None
-    ) -> tuple[np.ndarray, None, dict[str, float]]:
-        obs = self.task.collate([obs], self.device)
+    ) -> tuple[np.ndarray, Any, dict[str, float]]:
+        obs = self.task.collate([obs], device=self.device)
+
         with torch.no_grad():
-            action, _, stats = self.pi(obs, deterministic=deterministic)
-        return action.cpu().numpy()[0], None, stats
+            pi_output = self.pi(obs, state, deterministic=deterministic)
+
+        action = pi_output.action.cpu().numpy()[0]
+
+        return action, pi_output.state_solution, pi_output.stats
 
     @property
     def optimizers(self) -> list[torch.optim.Optimizer]:
