+++ tests/leap_c/torch/rl/test_buffer.py
@@ -7,9 +7,9 @@ from acados_template.acados_ocp_iterate import (
     AcadosOcpFlattenedBatchIterate,
     AcadosOcpFlattenedIterate,
 )
-from leap_c.collate import safe_collate_possible_nones
-from leap_c.acados.mpc import MpcParameter
-from leap_c.rl.buffer import ReplayBuffer
+from leap_c.torch.utils.collate import safe_collate_possible_nones
+from leap_c.ocp.acados.mpc import MpcParameter
+from leap_c.torch.rl.buffer import ReplayBuffer
 
 
 class ForNesting(NamedTuple):
